Project: rev 216  (#296)
Darwin -
* Project: rev 216
Note- almost finished still some buffer methods out of wack and I think some got named wrong

+++ api/src/main/java/net/runelite/api/Buffer.java
@@ -11,7 +11,7 @@ public interface Buffer extends Node
 
     byte readByteAdd();
     byte readByteNeg();
-    byte readUnsignedByteSub$api();
+    int readUnsignedByteSub$api();
 	int readUnsignedByte$api();
 	byte readByte$api();
 	int readUnsignedShort$api();
